From: Atari Ace [mailto:atari_ace@verizon.net] Sent: Saturday, December 13, 2008 5:07 PM To: submit@mamedev.org Cc: atariace@hotmail.com Subject: [patch] Add machine to allocgfx
Hi mamedev,

This patch eliminates the #include "deprecat.h" from drawgfx.h.  It
does so in a fashion similar to my recent tilemap patch, adding the
machine pointer to gfx_element, changing allocgfx to take a machine,
and then adjusting the internals to use the machine field as needed.
The changes outside of drawgfx.[ch] were done with the attached
script.

~aa