new clones -----
Automat (bootleg of Robocop) [Andrew Welburn, Philip Bennett, David Haywood]

this is a heavily modified bootleg of Robocop.  Phil hooked up some sound stuff, and I've got the graphics decoding sorted out.

the tilemap transparency handling is different to the original, as are the sprites, this will need a significant amount of work 

one of the sprite roms is bad, but it looks like it should be possible to reconstruct it correctly from the main set.