From: R. Reucher [rene.reucher@batcom-it.net] Sent: Friday, February 27, 2009 10:17 AM To: submit@mamedev.org Cc: qmc2-devel@lists.sourceforge.net Subject: Proposed patch for src/emu/audit.c
Dear MAME developers,

I've created the attached diff to src/emu/audit.c (0.129u5) to allow for 
reporting a ROM state of "correct" when a game/machine doesn't require any 
ROM files at all (currently it will be reported as "not found", which is 
wrong IMHO).

I know that this isn't really relevant for MAME, but it is for MESS -- so 
I hope you'll accept the patch.

Thanks, Rene
--
Rene Reucher
rene.reucher@batcom-it.net
http://www.batcom-it.net/

It is only people of small moral stature who have to stand on their dignity.