New clones ---------- Street Fighter II - The World Warrior (World 910214, TAB Austria bootleg)
From: Klaus Sommer, B.Sc [mailto:klaus@hotsummer.at] 
Sent: Tuesday, May 26, 2009 10:29 PM
To: submit@mamedev.org
Subject: E-Mail schreiben an: sf2ebbl.zip, sf2tabaustria.diff

hello!
 
we dumped a new bootleg of Street Fighter 2. One Differenz to the original set is that the logo's (capcom) are blanked.
 
Attached are the romset (Maskroms are the same as in the original set) and the DIFF...which is already verified from Haze!
 
Your's
TeamEurope