From: hoge hoge [mailto:c8cv@hotmail.com] Sent: Thursday, July 16, 2009 6:59 PM To: submit@mamedev.org Subject: namcos22 timecris crosshair fix
Hello,

Attached is a diff for 0132u5 that fixes a problem with the Time Crisis gun crosshair. 
The right and bottom parts of the screen couldn't be accessed, and you couldn't shoot 
off-screen. Off-screen shooting has a function in test mode, in the fix this is the 
bottom edge of the screen. This fix requires the user to (re)calibrate the gun via 
test mode.

Greets,
hap