From: David Haywood <neohaze@nildram.co.uk>
Date: Sun, 26 Jul 2009 09:31:24 
To: Aaron Giles<aaron@aarongiles.com>; Angelo Salese<salese_corp_ltd@email.it>
Subject: Space Walk?


I got the Software This guy to redump the 'Space Walk' rom which was 
obviously bad, and added it to the driver.

A second opinion is needed, maybe there is something else wrong because 
the gameplay doesn't appear to make any sense at all.  It's a similar 
game to clowns, but more often than not the character falls through the 
pad, and the pads jump and flip positions when they reach the middle of 
the screen.  Since this falls into the 'ultra rare' category I think it 
would be a good idea if somebody with a good knowledge of the CPU made 
absolutely sure that things are correct.


New not-working
===============
Space Walk [Siftware, David Haywood]